Create KASPar_plink.R
Using a sample file (first 6 columns of a PED file) and a MAP file as input this function creates a PED file from the cleaned and annotated LC480 genotyping data.  Can work on multiple input files at once as long as the rsid in the MAP file corresponds to the rsid of the file name.